Fork blue_balancer that does not depend on pink
The reason for this is that Bazel pip rules don't work for pinocchio
(because of the way it and its dependencies assume they share an
../openrobots.libs directory, which violates Bazel's hermeticity), and
therefore not for Pink which depends on it. We will be able to run the
blue balancer out of the box from Bazel, while asking users to ``pip
install pin-pink`` separately in order to use the pink balancer.