Reduce crouch height to avoid ankle bumps at max speed
These bumps are a combination of the steady-state ankle that the ankle
makes with the ground and the velocity at which Upkie crouches. The blue
controller crouches faster, therefore either the crouch height or
velocity should be adapted. This should avoid newcomers bumping into
this at their first simulation. (The issue is otherwise flagged, e.g. in
the YouTube video. To be addressed in next ankle redesign.)