Remove unit tests for now
Continuous integration with PyTorch on GitHub Actions is not
straightforward. The current issue is:

```
OSError: libcublas.so.11: cannot open shared object file: No such file or directory
```

Despite the corresponding library being in our
``requirements_lock.txt``. Switching to the following, in that file,
could be an option:

```
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.13.1+cpu
```

But then users would get the CPU rather than GPU version of PyTorch,
paying a price that's only needed for the CI. Unit testing this agent is
not top priority for now, so let's stash this issue.